requirements.txt
Flask==2.3.3
google-generativeai==0.7.2
FPDF==1.7.2
python-docx==0.8.11
